+++ DemoApp/StreamChat/Components/DemoChatChannelListErrorView.swift
@@ -1,0 +1,20 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+import StreamChatUI
+import UIKit
+
+class DemoChatChannelListErrorView: ChatChannelListErrorView {
+    override func show() {
+        UIView.animate(withDuration: 0.5) {
+            self.isHidden = false
+        }
+    }
+
+    override func hide() {
+        UIView.animate(withDuration: 0.5) {
+            self.isHidden = true
+        }
+    }
+}