+++ Sources/StreamChat/Utils/StreamTimer/ScheduledStreamTimer.swift
@@ -5,17 +5,22 @@
 import Foundation
 
 public class ScheduledStreamTimer: StreamTimer {
-    let interval: TimeInterval
     var runLoop = RunLoop.current
     var timer: Foundation.Timer?
     public var onChange: (() -> Void)?
 
-    public var isRunning: Bool {
-        timer?.isValid ?? false
+    let interval: TimeInterval
+    let fireOnStart: Bool
+    let repeats: Bool
+
+    public init(interval: TimeInterval, fireOnStart: Bool = true, repeats: Bool = true) {
+        self.interval = interval
+        self.fireOnStart = fireOnStart
+        self.repeats = repeats
     }
 
-    public init(interval: TimeInterval) {
-        self.interval = interval
+    public var isRunning: Bool {
+        timer?.isValid ?? false
     }
 
     public func start() {
@@ -23,12 +28,14 @@ public class ScheduledStreamTimer: StreamTimer {
 
         timer = Foundation.Timer.scheduledTimer(
             withTimeInterval: interval,
-            repeats: true
+            repeats: repeats
         ) { _ in
             self.onChange?()
         }
         runLoop.add(timer!, forMode: .common)
-        timer?.fire()
+        if fireOnStart {
+            timer?.fire()
+        }
     }
 
     public func stop() {