+++ Sources/StreamChat/WebSocketClient/WebSocketClient.swift
@@ -142,6 +142,17 @@ class WebSocketClient {
             eventsBatcher.processImmediately(completion: completion)
         }
     }
+
+    func timeout() {
+        let previousState = connectionState
+        connectionState = .disconnected(source: .timeout(from: previousState))
+        engineQueue.async { [engine, eventsBatcher] in
+            engine?.disconnect()
+
+            eventsBatcher.processImmediately {}
+        }
+        log.error("Connection timed out. `\(connectionState)", subsystems: .webSocket)
+    }
 }
 
 protocol ConnectionStateDelegate: AnyObject {